api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snap2recipe-api"
version = "1.0.0"
description = "FastAPI backend for Snap2Recipe ingredient detection and recipe suggestion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pillow>=10.1.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "opencv-python-headless>=4.8.1",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "rank-bm25>=0.2.2",
    "nltk>=3.8.1",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.11.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["api"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
